release candidate 5

+++ scripts/game_delete_data/game_delete_data.gml
@@ -3,5 +3,8 @@
 /// @param {Real} _slot	The slot index of the saved data to delete.
 function game_delete_data(_slot)
 {
-	file_delete("save" + string(_slot) + ".bin");
+	if _slot != undefined
+	{
+		file_delete("save" + string(_slot) + ".bin");
+	}
 }
